requirements.txt
-r requirements/base.txt

asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-on-heroku==1.1.2
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
